Add Grading policy app.
