+++ common/lib/xmodule/xmodule/course_module.py
@@ -13,7 +13,6 @@ from lazy import lazy
 
 from xmodule import course_metadata_utils
 from xmodule.course_metadata_utils import DEFAULT_START_DATE
-from xmodule.exceptions import UndefinedContext
 from xmodule.seq_module import SequenceDescriptor, SequenceModule
 from xmodule.graders import grader_from_conf
 from xmodule.tabs import CourseTabList
@@ -24,6 +23,9 @@ from xblock.core import XBlock
 from xblock.fields import Scope, List, String, Dict, Boolean, Integer, Float
 from .fields import Date
 from django.utils.timezone import UTC
+from django.conf import settings
+from stevedore.extension import ExtensionManager
+
 
 
 log = logging.getLogger(__name__)
@@ -168,6 +170,11 @@ class TextbookList(List):
             else:
                 continue
         return json_data
+
+
+class GradingPolicyError(Exception):
+    """An error occurred in the Grading Policy App. """
+    pass
 
 
 class CourseFields(object):
@@ -1288,6 +1295,16 @@ class CourseDescriptor(CourseFields, SequenceDescriptor, LicenseMixin):
 
         return announcement, start, now
 
+
+    @lazy
+    def grading(self):
+        name = settings.GRADING_TYPE
+        extension = ExtensionManager(namespace='openedx.grading_policy')
+        try:
+            return extension[name].plugin
+        except KeyError:
+            raise GradingPolicyError("Unrecognized grading strategy `{0}`".format(name))
+
     @lazy
     def grading_context(self):
         """
@@ -1310,51 +1327,8 @@ class CourseDescriptor(CourseFields, SequenceDescriptor, LicenseMixin):
             effect grading a student. This is used to efficiently fetch
             all the xmodule state for a FieldDataCache without walking
             the descriptor tree again.
-
-
         """
-        # If this descriptor has been bound to a student, return the corresponding
-        # XModule. If not, just use the descriptor itself
-        try:
-            module = getattr(self, '_xmodule', None)
-            if not module:
-                module = self
-        except UndefinedContext:
-            module = self
-
-        def possibly_scored(usage_key):
-            """Can this XBlock type can have a score or children?"""
-            return usage_key.block_type in self.block_types_affecting_grading
-
-        all_descriptors = []
-        graded_sections = {}
-
-        def yield_descriptor_descendents(module_descriptor):
-            for child in module_descriptor.get_children(usage_key_filter=possibly_scored):
-                yield child
-                for module_descriptor in yield_descriptor_descendents(child):
-                    yield module_descriptor
-
-        for chapter in self.get_children():
-            for section in chapter.get_children():
-                if section.graded:
-                    xmoduledescriptors = list(yield_descriptor_descendents(section))
-                    xmoduledescriptors.append(section)
-
-                    # The xmoduledescriptors included here are only the ones that have scores.
-                    section_description = {
-                        'section_descriptor': section,
-                        'xmoduledescriptors': [child for child in xmoduledescriptors if child.has_score]
-                    }
-
-                    section_format = section.format if section.format is not None else ''
-                    graded_sections[section_format] = graded_sections.get(section_format, []) + [section_description]
-
-                    all_descriptors.extend(xmoduledescriptors)
-                    all_descriptors.append(section)
-
-        return {'graded_sections': graded_sections,
-                'all_descriptors': all_descriptors, }
+        return self.grading.grading_context(self)
 
     @lazy
     def block_types_affecting_grading(self):
